Move email helper into root directory
